requirements.txt
numpy~=1.26.4
torchvision~=0.20.1+cu121
ultralytics~=8.3.59
glfw~=2.8.0
pyopengl~=3.1.7
imageio~=2.36.1
tqdm~=4.67.1
colorama~=0.4.6
ffmpeg~=7.1.0
ffmpeg-python~=0.2.0
pycuda~=2024.1
opencv-python~=4.10.0.84
scipy~=1.15.1
matplotlib~=3.10.0
simplification~=0.7.13
torch~=2.5.1+cu121